Halfway through refactor
